Fix to show separate tail / cycle color in /rgb & /colors preview RECONNECT_TO_LAST_SERVER_ON_START -  auto reconnect to last server on start Flipped color previews to show correct color in menu for boxes & cycle / tail refaactor

+++ src/engine/eChatCommands.h
@@ -147,7 +147,7 @@ class ColorsCommand : public ChatCommand
 public:
     ColorsCommand() : ChatCommand("ColorsCommand") {}
 
-    static tColoredString cycleColorPreview(int r, int g, int b);
+    static tColoredString cycleColorPreview(REAL r, REAL g, REAL b);
 
     static tColoredString localPlayerPreview(ePlayer *local_p);
